refactored skat to stats
